Updated plot_resolution with improved plots and correct half-angles
There is still an issue in terms of the outputs - the numbers seem lower than expected, particularly the term for chromatic aberration giving a solution to the order of E-31. I would estimate these should all be on the order of 2nm. Let me know if you see any issues and I will take a look myself too